UI design  -Math-magicians
Create a directory called components
Inside components, create a new Calculator.js file.
In Calculator.js, create a React component that matches this design
Don't add any extra design.
Develop just the UI of the calculator; the math logic to make it work will be added in the next project
Use class based components (All components should be class-based)
Once the Calculator is ready, import it and use it in your main component (App.js).
Test it in the browser; it should render without problems.